3  -- API_BDD_carter_cash/API/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
DateTime==5.4
dnspython==2.6.1
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
orjson==3.9.15
passlib==1.7.4
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.2
pydantic-extra-types==2.5.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyodbc==5.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
ujson==5.9.0
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
zope.interface==6.2